Added `page_title_add_segment` Liquid tag to allow changing the page title without rendering it.
